Create Modified version including ALOS DSM
Multi-Scale Relief Model script originally published as: Orengo, H.A. and Petrie, C.A. 2018. 'Multi-Scale Relief Model (MSRM): a new algorithm for the analysis of subtle topographic change in digital elevation models' Earth Surface Processes and Landforms, 43(6): 1361-1369.
This version differs from the original as it incorporates ALOS30 DSM as a variable (much improved since the original paper was published) and some other minor changes that do not affect the workings of the original script.